pyproject.toml
[project]
name = "imageconverter"
version = "0.1.0"
description = ""
authors = [
    {name = "Nil Mallick",email = "mallick.nilotpal@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pillow (>=11.1.0,<12.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
